Move test helper functions
