+++ slam-libraries/viam-cartographer/src/io/file_handler_test.cc
@@ -1,4 +1,5 @@
 #include "file_handler.h"
+#include "../utils/test_helpers.h"
 
 #include <boost/filesystem.hpp>
 #include <boost/filesystem/fstream.hpp>
@@ -11,65 +12,6 @@ namespace io {
 namespace {
 
 BOOST_AUTO_TEST_SUITE(file_handler)
-
-// NOTE: When using this function, make sure to remove the temporary directory
-// and its contents once done processing the files by calling the function
-// removeTmpDirectory(tmpdir).
-boost::filesystem::path createTmpDirectoryAndAddFiles(std::vector<std::string> data_files,
-                                                std::vector<std::string> map_files) {
-    // Create a unique path in the temp directory and add the files
-    boost::filesystem::path tmpdir = boost::filesystem::temp_directory_path() /
-                                     boost::filesystem::unique_path();
-    bool ok = boost::filesystem::create_directory(tmpdir);
-    if (!ok) {
-        throw std::runtime_error("could not create directory: " +
-                                 tmpdir.string());
-    }
-    // Create the "config" subdirectory
-    boost::filesystem::path tmpdir_config = tmpdir / "config";
-    ok = boost::filesystem::create_directory(tmpdir_config);
-    if (!ok) {
-        boost::filesystem::remove_all(tmpdir);
-        throw std::runtime_error("could not create directory: " +
-                                 tmpdir_config.string());
-    }
-
-    // Create the "data" subdirectory
-    boost::filesystem::path tmpdir_data = tmpdir / "data";
-    ok = boost::filesystem::create_directory(tmpdir_data);
-    if (!ok) {
-        boost::filesystem::remove_all(tmpdir);
-        throw std::runtime_error("could not create directory: " +
-                                 tmpdir_data.string());
-    }
-
-    // Create the "map" subdirectory
-    boost::filesystem::path tmpdir_map = tmpdir / "map";
-    ok = boost::filesystem::create_directory(tmpdir_map);
-    if (!ok) {
-        boost::filesystem::remove_all(tmpdir);
-        throw std::runtime_error("could not create directory: " +
-                                 tmpdir_map.string());
-    }
-
-    // Add data files to "data" subdirectory
-    for (std::string file : data_files) {
-        boost::filesystem::ofstream ofs(tmpdir_data / file);
-        ofs.close();
-    }
-
-    // Add map files to "map" subdirectory
-    for (std::string file : map_files) {
-        boost::filesystem::ofstream ofs(tmpdir_map / file);
-        ofs.close();
-    }
-
-    return tmpdir;
-}
-
-void removeTmpDirectory(boost::filesystem::path tmpdir) {
-    boost::filesystem::remove_all(tmpdir);
-}
 
 BOOST_AUTO_TEST_CASE(MakeFilenameWithTimestamp_success) {
     std::string path_to_dir = "path_to_dir";
@@ -101,7 +43,7 @@ BOOST_AUTO_TEST_CASE(ListSortedFilesInDirectory_success) {
         "rplidar_data_2022-01-01T01:00:00.0003Z.pcd"};
     std::vector<std::string> map_files{};
     // Create a unique path in the temp directory and add the files
-    boost::filesystem::path tmpdir = createTmpDirectoryAndAddFiles(data_files, map_files);
+    boost::filesystem::path tmpdir = utils::createTmpDirectoryAndAddFiles(data_files, map_files);
     
     // List the sorted files in the directory
     std::vector<std::string> listed_data_files =
@@ -113,7 +55,7 @@ BOOST_AUTO_TEST_CASE(ListSortedFilesInDirectory_success) {
         BOOST_TEST(data_files.at(i).compare(listed_data_files.at(i)));
     }
     // Remove the temporary directory and its contents
-    removeTmpDirectory(tmpdir);
+    utils::removeTmpDirectory(tmpdir);
 }
 
 // test RemoveFile
@@ -126,7 +68,7 @@ BOOST_AUTO_TEST_CASE(RemoveFile_success) {
         "rplidar_data_2022-01-01T01:00:00.0003Z.pcd"};
     std::vector<std::string> map_files{};
     // Create a unique path in the temp directory and add the files
-    boost::filesystem::path tmpdir = createTmpDirectoryAndAddFiles(data_files, map_files);
+    boost::filesystem::path tmpdir = utils::createTmpDirectoryAndAddFiles(data_files, map_files);
 
     // Remove a file
     int file_num = 1;
@@ -142,7 +84,7 @@ BOOST_AUTO_TEST_CASE(RemoveFile_success) {
         BOOST_TEST(data_files.at(i).compare(listed_data_files.at(i)));
     }
     // Remove the temporary directory and its contents
-    removeTmpDirectory(tmpdir);
+    utils::removeTmpDirectory(tmpdir);
 }
 
 BOOST_AUTO_TEST_CASE(TimedPointCloudDataFromPCDBuilder_success) {
@@ -198,7 +140,7 @@ BOOST_AUTO_TEST_CASE(TimedPointCloudDataFromPCDBuilder_success) {
     }
 
     // Remove the temporary directory and its contents
-    boost::filesystem::remove_all(tmpdir);
+    utils::removeTmpDirectory(tmpdir);
 }
 
 BOOST_AUTO_TEST_CASE(ReadTimeFromFilename_success) {
